rgblight.h no longer changes behavior when included >1 time
Headers should never change behavior dependent on the number of times
they're included. This fixes that defect.
